pyproject.toml
[tool.poetry]
name = "aerith-ingestion"
version = "0.1.0"
description = "A tool to help you manage your todoist tasks"
authors = ["Albert J Orlando <albertjorlando@gmail.com>"]
readme = "README.md"

packages = [
    { include = "aerith_ingestion", from = "src" }
]

[tool.poetry.dependencies]
chromadb = "^0.4.24"
jsonschema = "^4.23.0"
llama-index = "^0.10.14"
llama-index-vector-stores-chroma = "^0.1.3"
loguru = "^0.7.3"
pydantic-ai = {extras = ["logfire"], version = "^0.0.18"}
pydantic-settings = "^2.2.1"
python = ">=3.11.5,<3.12"
python-dotenv = "^1.0.1"
pytest-asyncio = "^0.25.2"
todoist-api-python = "^2.1.3"
langchain = "^0.3.14"
langchain-openai = "^0.3.0"
openai = "^1.59.7"
sentence-transformers = "^3.3.1"
plotly = "^5.24.1"
scikit-learn = "^1.6.1"
umap-learn = "^0.5.7"
py2puml = "^0.9.1"
pymermaider = "^0.1.4"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
click = "^8.1.7"
faiss-cpu = "^1.9.0.post1"
crawl4ai = "^0.4.247"
langchain-community = "^0.3.14"
mdformat = "^0.7.21"

[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
flake8 = "^6.0.0"
isort = "^5.13.2"
flake8-isort = "^6.1.1"
pycln = "^2.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
aerith = "aerith_ingestion.cli:cli"

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"

[tool.pycln]
all = true
